+++ Source/DynamicScene/DynamicBillboardVisualizer.js
@@ -351,6 +351,11 @@ define([
         if (defined(property)) {
             billboard.setHeight(property.getValue(time));
         }
+
+        property = dynamicBillboard._nearFarScalar;
+        if (defined(property)) {
+            billboard.setScaleByDistance(property.getValue(time));
+        }
     }
 
     DynamicBillboardVisualizer.prototype._onObjectsRemoved = function(dynamicObjectCollection, added, dynamicObjects) {