+++ Specs/DynamicScene/DynamicBillboardVisualizerSpec.js
@@ -11,6 +11,7 @@ defineSuite([
              'Core/Cartesian2',
              'Core/Cartesian3',
              'Core/Color',
+             'Core/NearFarScalar',
              'Scene/Scene',
              'Scene/BillboardCollection',
              'Scene/HorizontalOrigin',
@@ -27,6 +28,7 @@ defineSuite([
               Cartesian2,
               Cartesian3,
               Color,
+              NearFarScalar,
               Scene,
               BillboardCollection,
               HorizontalOrigin,
@@ -151,6 +153,7 @@ defineSuite([
             billboard.pixelOffset = new ConstantProperty(new Cartesian2(3, 2));
             billboard.width = new ConstantProperty(15);
             billboard.height = new ConstantProperty(5);
+            billboard.nearFarScalar = new ConstantProperty(new NearFarScalar());
 
             visualizer.update(time);
 
@@ -171,6 +174,7 @@ defineSuite([
                     expect(bb.getVerticalOrigin()).toEqual(testObject.billboard.verticalOrigin.getValue(time));
                     expect(bb.getWidth()).toEqual(testObject.billboard.width.getValue(time));
                     expect(bb.getHeight()).toEqual(testObject.billboard.height.getValue(time));
+                    expect(bb.getScaleByDistance()).toEqual(testObject.billboard.nearFarScalar.getValue(time));
                 }
                 return bb.getShow(); //true once the image is loaded.
             });
@@ -190,6 +194,7 @@ defineSuite([
             billboard.pixelOffset = new ConstantProperty(new Cartesian2(2, 3));
             billboard.width = new ConstantProperty(17);
             billboard.height = new ConstantProperty(12);
+            billboard.nearFarScalar = new ConstantProperty(new NearFarScalar());
 
             waitsFor(function() {
                 visualizer.update(time);
@@ -206,6 +211,7 @@ defineSuite([
                     expect(bb.getPixelOffset()).toEqual(testObject.billboard.pixelOffset.getValue(time));
                     expect(bb.getWidth()).toEqual(testObject.billboard.width.getValue(time));
                     expect(bb.getHeight()).toEqual(testObject.billboard.height.getValue(time));
+                    expect(bb.getScaleByDistance()).toEqual(testObject.billboard.nearFarScalar.getValue(time));
                 }
                 return imageReady;
             });